deck & nav

+++ assets/core/components/subscriptionButton.vue
@@ -1,35 +1,44 @@
 <template>
     <p-modal ref="modal">
-        <p-subscription-form @done="$refs.modal.hide()" class="box" :context="context"
-            :initial="subscription"></p-subscription-form>
-    </p-modal>
-    <div class="dropdown is-hoverable is-right" v-if="!subscription">
-        <div class="dropdown-trigger">
-            <button class="button is-link" @click="subscribe"
-                    :title="`Subscribe as ${roles[context.subscription_default_role]}`">
-                <span class="icon"><i class="mdi mdi-account-multiple"></i></span>
-                <span>Subscribe</span>
-            </button>
+        <div class="box">
+            <h2 class="title is-2">
+                <span v-if="subscription">Edit subscription</span>
+                <span v-else>Subscribe to {{ context && context.title }}</span>
+            </h2>
+            <p-subscription-form @done="$refs.modal.hide()" :context="context"
+                :initial="subscription"></p-subscription-form>
         </div>
-        <div class="dropdown-menu" role="menu">
-            <div class="dropdown-content">
-                <a class="dropdown-item" @click="edit">Subscribe...</a>
+    </p-modal>
+    <div class="dropdown is-hoverable is-right">
+        <div class="dropdown-trigger">
+            <div class="field has-addons">
+                <div class="control">
+                    <button v-if="!subscription" class="button is-link" @click="subscribe"
+                            :title="`Subscribe as ${roles[context.subscription_default_role]}`">
+                        <span class="icon"><i class="mdi mdi-account-multiple"></i></span>
+                        <span>Subscribe</span>
+                    </button>
+                    <button v-else-if="subscription.isSubscribed" class="button is-info" @click="edit">
+                        <span class="icon"><i class="mdi mdi-account-multiple"></i></span>
+                        <span>{{ roles[context.role.access].name }}</span>
+                    </button>
+                    <button v-else class="button is-info" @click="edit">
+                        <span class="icon"><i class="mdi mdi-account-question"></i></span>
+                        <span>Request sent</span>
+                    </button>
+                </div>
+                <div class="control">
+                    <button class="button is-white">
+                        {{ context.n_subscriptions }}
+                    </button>
+                </div>
             </div>
         </div>
-    </div>
-    <div class="dropdown is-hoverable is-right" v-else>
-        <div class="dropdown-trigger">
-            <button class="button" @click="edit" v-if="subscription.isSubscribed">
-                <span class="icon"><i class="mdi mdi-account-multiple"></i></span>
-                <span>{{ roles[context.role.access].name }}</span>
-            </button>
-            <button class="button is-info" @click="edit" v-else>
-                <span class="icon"><i class="mdi mdi-account-multiple"></i></span>
-                <span>Request sent</span>
-            </button>
-        </div>
         <div class="dropdown-menu" role="menu">
-            <div class="dropdown-content">
+            <div v-if="!subscription" class="dropdown-content">
+                <a class="dropdown-item" @click="edit">Subscribe...</a>
+            </div>
+            <div v-else class="dropdown-content">
                 <a class="dropdown-item" @click="unsubscribe">Unsubscribe</a>
             </div>
         </div>
@@ -37,7 +46,9 @@
 </template>
 <script>
 import PSubscriptionForm from './subscriptionForm'
+import PSubscriptionList from './subscriptionList'
 import PModal from './modal'
+
 
 export default {
     inject: ['roles'],
@@ -59,11 +70,20 @@ export default {
             this.$refs.modal.show()
         },
 
+        subscribe() {
+            return new this.$root.Subscription({
+                context_id: this.context.$id,
+                access: this.context.subscription_default_access,
+                role: this.context.subscription_default_role,
+            }).save()
+        },
+
         unsubscribe() {
-            this.subscription.delete()
+            confirm(`Unsubscribe from ${this.context.title}?`) &&
+                this.subscription.delete()
         },
     },
 
-    components: { PSubscriptionForm, PModal },
+    components: { PSubscriptionForm, PSubscriptionList, PModal },
 }
 </script>
